:recycle: refactor : Game 도메인의 Redis를 Repository layer로 분리

+++ BE/the-knight/src/main/java/com/a301/theknight/domain/game/dto/playing/request/GameOrderRequest.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class GameOrderRequest {
     private int orderNumber;
 
-    public boolean validate(long capacity) {
+    public boolean validate(int capacity) {
         return orderNumber >= 1 && orderNumber <= capacity;
     }
 }
